+++ packages/firebase_analytics/firebase_analytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+* Make the pedantic dev_dependency explicit.
+
 ## 1.0.1
 
 * Remove the deprecated `author:` field from pubspec.yaml