+++ packages/firebase_database/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.3
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 3.1.2
+
+* Make the pedantic dev_dependency explicit.
+
 ## 3.1.1
 
 * Remove the deprecated `author:` field from pubspec.yaml