+++ packages/firebase_ml_vision/CHANGELOG.md
@@ -3,6 +3,18 @@
 * **Breaking Change** Update Android firebase-ml-vision dependencies to latest.
 * Add `rawBytes` property to `Barcode`.
 
+## 0.9.3+8
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 0.9.3+7
+
+* Make the pedantic dev_dependency explicit.
+
+## 0.9.3+6
+
+* Replace deprecated text theme fields in example app.
+
 ## 0.9.3+5
 
 * Remove the deprecated `author:` field from pubspec.yaml