+++ packages/firebase_remote_config/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0+3
+
+* Replace deprecated `getFlutterEngine` call on Android.
+
+## 0.3.0+2
+
+* Make the pedantic dev_dependency explicit.
+
 ## 0.3.0+1
 
 * Remove the deprecated `author:` field from pubspec.yaml